requirements.txt
langgraph>=0.2.40
langchain-core>=0.3.10
langchain-groq>=0.1.5
groq>=0.9.0
python-dotenv>=1.0.1

